refactor: added default implementation for observe and updated atom & atom:undo & reaction & access_state
