+++ src/reactive_state_access/state_access.rs
@@ -178,26 +178,6 @@ impl<T> Observable<T> for StateAccess<T>
 where
     T: 'static,
 {
-    fn observe(&self) -> T
-    where
-        T: Clone,
-    {
-        let id = crate::store::StorageKey::TopoKey(self.id);
-        let context = illicit::get::<RefCell<ReactiveContext>>().expect(
-            "No #[reaction] context found, are you sure you are in one? I.e. does the current \
-             function have a #[reaction] tag?",
-        );
-        context.borrow_mut().reactive_state_accessors.push(id);
-
-        STORE.with(|store_refcell| {
-            store_refcell
-                .borrow_mut()
-                .add_dependency(&id, &context.borrow().key);
-        });
-
-        clone_reactive_state_with_id::<T>(id).unwrap()
-    }
-
     #[topo::nested]
     fn observe_update(&self) -> (Option<T>, T)
     where
